Organize preview endpoints for data and instance
