+++ backend/src/Designer/Controllers/PreviewController.cs
@@ -7,7 +7,6 @@ using System.Text.Json.Nodes;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Web;
-using Altinn.App.Core.Internal.Process.Elements;
 using Altinn.Platform.Profile.Models;
 using Altinn.Platform.Register.Enums;
 using Altinn.Platform.Register.Models;
@@ -30,49 +29,39 @@ namespace Altinn.Studio.Designer.Controllers
     /// <summary>
     /// Controller containing all actions related to preview - still under development
     /// </summary>
+    /// <remarks>
+    /// Initializes a new instance of the <see cref="PreviewController"/> class.
+    /// </remarks>
+    /// <param name="httpContextAccessor"></param>
+    /// <param name="altinnGitRepositoryFactory">IAltinnGitRepositoryFactory</param>
+    /// <param name="schemaModelService">Schema Model Service</param>
+    /// <param name="previewService">Preview Service</param>
+    /// <param name="textsService">Texts Service</param>
+    /// <param name="appDevelopmentService">App Development Service</param>
+    /// Factory class that knows how to create types of <see cref="AltinnGitRepository"/>
     [Authorize]
     [AutoValidateAntiforgeryToken]
     // Uses regex to not match on designer since the call from frontend to get the iframe for app-frontend,
     // `designer/html/preview.html`, will match on Image-endpoint which is a fetch-all route
     [Route("{org:regex(^(?!designer))}/{app:regex(^(?!datamodels$)[[a-z]][[a-z0-9-]]{{1,28}}[[a-z0-9]]$)}")]
-    public class PreviewController : Controller
+    public class PreviewController(IHttpContextAccessor httpContextAccessor,
+        IAltinnGitRepositoryFactory altinnGitRepositoryFactory,
+        ISchemaModelService schemaModelService,
+        IPreviewService previewService,
+        ITextsService textsService,
+        IAppDevelopmentService appDevelopmentService) : Controller
     {
-        private readonly IHttpContextAccessor _httpContextAccessor;
-        private readonly IAltinnGitRepositoryFactory _altinnGitRepositoryFactory;
-        private readonly ISchemaModelService _schemaModelService;
-        private readonly IPreviewService _previewService;
-        private readonly ITextsService _textsService;
-        private readonly IAppDevelopmentService _appDevelopmentService;
+        private readonly IHttpContextAccessor _httpContextAccessor = httpContextAccessor;
+        private readonly IAltinnGitRepositoryFactory _altinnGitRepositoryFactory = altinnGitRepositoryFactory;
+        private readonly ISchemaModelService _schemaModelService = schemaModelService;
+        private readonly IPreviewService _previewService = previewService;
+        private readonly ITextsService _textsService = textsService;
+        private readonly IAppDevelopmentService _appDevelopmentService = appDevelopmentService;
 
         // This value will be overridden to act as the task number for apps that use layout sets
         private const int PartyId = 51001;
         private const string MINIMUM_NUGET_VERSION = "8.0.0.0";
         private const int MINIMUM_PREVIEW_NUGET_VERSION = 15;
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="PreviewController"/> class.
-        /// </summary>
-        /// <param name="httpContextAccessor"></param>
-        /// <param name="altinnGitRepositoryFactory">IAltinnGitRepositoryFactory</param>
-        /// <param name="schemaModelService">Schema Model Service</param>
-        /// <param name="previewService">Preview Service</param>
-        /// <param name="textsService">Texts Service</param>
-        /// <param name="appDevelopmentService">App Development Service</param>
-        /// Factory class that knows how to create types of <see cref="AltinnGitRepository"/>
-        public PreviewController(IHttpContextAccessor httpContextAccessor,
-            IAltinnGitRepositoryFactory altinnGitRepositoryFactory,
-            ISchemaModelService schemaModelService,
-            IPreviewService previewService,
-            ITextsService textsService,
-            IAppDevelopmentService appDevelopmentService)
-        {
-            _httpContextAccessor = httpContextAccessor;
-            _altinnGitRepositoryFactory = altinnGitRepositoryFactory;
-            _schemaModelService = schemaModelService;
-            _previewService = previewService;
-            _textsService = textsService;
-            _appDevelopmentService = appDevelopmentService;
-        }
 
         /// <summary>
         /// Default action for the preview.
@@ -265,20 +254,6 @@ namespace Altinn.Studio.Designer.Controllers
         }
 
         /// <summary>
-        /// Action for getting a response from v1/data/anonymous
-        /// </summary>
-        /// <param name="org">Unique identifier of the organisation responsible for the app.</param>
-        /// <param name="app">Application identifier which is unique within an organisation.</param>
-        /// <returns>Empty object</returns>
-        [HttpGet]
-        [Route("api/v1/data/anonymous")]
-        public IActionResult Anonymous(string org, string app)
-        {
-            string user = "{}";
-            return Content(user);
-        }
-
-        /// <summary>
         /// Action for responding to keepAlive
         /// </summary>
         /// <param name="org">Unique identifier of the organisation responsible for the app.</param>
@@ -399,25 +374,6 @@ namespace Altinn.Studio.Designer.Controllers
         }
 
         /// <summary>
-        /// Action for creating the mocked instance object
-        /// </summary>
-        /// <param name="org">Unique identifier of the organisation responsible for the app.</param>
-        /// <param name="app">Application identifier which is unique within an organisation.</param>
-        /// <param name="instanceOwnerPartyId"></param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/> that observes if operation is cancelled.</param>
-        /// <returns>The mocked instance object</returns>
-        [HttpPost]
-        [Route("instances")]
-        public async Task<ActionResult<Instance>> Instances(string org, string app, [FromQuery] int? instanceOwnerPartyId, CancellationToken cancellationToken)
-        {
-            string developer = AuthenticationHelper.GetDeveloperUserName(_httpContextAccessor.HttpContext);
-            string refererHeader = Request.Headers["Referer"];
-            string layoutSetName = GetSelectedLayoutSetInEditorFromRefererHeader(refererHeader);
-            Instance mockInstance = await _previewService.GetMockInstance(org, app, developer, instanceOwnerPartyId, layoutSetName, cancellationToken);
-            return Ok(mockInstance);
-        }
-
-        /// <summary>
         /// Action for getting the mocked instance id
         /// </summary>
         /// <param name="org">Unique identifier of the organisation responsible for the app.</param>
@@ -433,198 +389,6 @@ namespace Altinn.Studio.Designer.Controllers
             string layoutSetName = GetSelectedLayoutSetInEditorFromRefererHeader(refererHeader);
             Instance mockInstance = await _previewService.GetMockInstance(org, app, developer, PartyId, layoutSetName, cancellationToken);
             return Ok(mockInstance.Id);
-        }
-
-        /// <summary>
-        /// Action for getting the json schema for the datamodel for the default data task test-datatask-id
-        /// </summary>
-        /// <param name="org">Unique identifier of the organisation responsible for the app.</param>
-        /// <param name="app">Application identifier which is unique within an organisation.</param>
-        /// <param name="partyId">party id</param>
-        /// <param name="instanceGuid">instance</param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/> that observes if operation is cancelled.</param>
-        /// <returns>Json schema for datamodel for the current task</returns>
-        [HttpGet]
-        [Route("instances/{partyId}/{instanceGuid}/data/" + PreviewService.MockDataTaskId)]
-        public async Task<ActionResult> GetFormData(string org, string app, [FromRoute] int partyId, [FromRoute] string instanceGuid, CancellationToken cancellationToken)
-        {
-            string developer = AuthenticationHelper.GetDeveloperUserName(_httpContextAccessor.HttpContext);
-            string refererHeader = Request.Headers["Referer"];
-            string layoutSetName = GetSelectedLayoutSetInEditorFromRefererHeader(refererHeader);
-            DataType dataType = await _previewService.GetDataTypeForLayoutSetName(org, app, developer, layoutSetName, cancellationToken);
-            // For apps that does not have a datamodel
-            if (dataType == null)
-            {
-                Instance mockInstance = await _previewService.GetMockInstance(org, app, developer, partyId, layoutSetName, cancellationToken);
-                return Ok(mockInstance.Id);
-            }
-            string modelPath = $"/App/models/{dataType.Id}.schema.json";
-            string decodedPath = Uri.UnescapeDataString(modelPath);
-            string formData = await _schemaModelService.GetSchema(AltinnRepoEditingContext.FromOrgRepoDeveloper(org, app, developer), decodedPath, cancellationToken);
-            return Ok(formData);
-        }
-
-        /// <summary>
-        /// Action for updating the json schema for the datamodel for the current data task in the process
-        /// </summary>
-        /// <param name="org">Unique identifier of the organisation responsible for the app.</param>
-        /// <param name="app">Application identifier which is unique within an organisation.</param>
-        /// <param name="partyId"></param>
-        /// <param name="instanceGuid"></param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/> that observes if operation is cancelled.</param>
-        /// <returns>Json schema for datamodel for the current data task in the process</returns>
-        [HttpPut]
-        [Route("instances/{partyId}/{instanceGuid}/data/" + PreviewService.MockDataTaskId)]
-        public async Task<ActionResult> UpdateFormData(string org, string app, [FromRoute] int partyId, [FromRoute] string instanceGuid, CancellationToken cancellationToken)
-        {
-            return await GetFormData(org, app, partyId, instanceGuid, cancellationToken);
-        }
-
-        /// <summary>
-        /// Action for mocking upload of an attachment to an attachment component
-        /// </summary>
-        /// <param name="dataType">Id of the attachment component in application metadata</param>
-        /// <returns>A 201 Created response with a mocked data element</returns>
-        [HttpPost]
-        [Route("instances/{partyId}/{instanceGuid}/data")]
-        public ActionResult PostAttachment([FromQuery] string dataType)
-        {
-            // This guid will be the unique id of the uploaded attachment
-            Guid guid = Guid.NewGuid();
-            DataElement dataElement = new() { Id = guid.ToString() };
-            return Created("link-to-app-placeholder", dataElement);
-        }
-
-        /// <summary>
-        /// Action for mocking deleting an uploaded attachment to an attachment component
-        /// </summary>
-        /// <param name="dataTypeId">Id of the attachment in application metadata</param>
-        /// <returns>Ok</returns>
-        [HttpDelete]
-        [Route("instances/{partyId}/{instanceGuid}/data/{dataTypeId}")]
-        public ActionResult DeleteAttachment([FromRoute] string dataTypeId)
-        {
-            return Ok();
-        }
-
-        /// <summary>
-        /// Action for mocking updating tags for an attachment component in the datamodel
-        /// </summary>
-        /// <param name="tag">The specific tag from the code list chosen for the attachment</param>
-        /// <returns>Ok</returns>
-        [HttpPost]
-        [Route("instances/{partyId}/{instanceGuid}/data/{dataTypeId}/tags")]
-        public ActionResult UpdateTagsForAttachment([FromBody] string tag)
-        {
-            return Created("link-to-app-placeholder", tag);
-        }
-
-        /// <summary>
-        /// Action for getting a mocked response for the current task connected to the instance
-        /// </summary>
-        /// <returns>The processState</returns>
-        [HttpGet]
-        [Route("instances/{partyId}/{instanceGuId}/process")]
-        public async Task<ActionResult<AppProcessState>> Process(string org, string app, [FromRoute] int partyId, CancellationToken cancellationToken)
-        {
-            string developer = AuthenticationHelper.GetDeveloperUserName(_httpContextAccessor.HttpContext);
-            string refererHeader = Request.Headers["Referer"];
-            string layoutSetName = GetSelectedLayoutSetInEditorFromRefererHeader(refererHeader);
-            Instance mockInstance = await _previewService.GetMockInstance(org, app, developer, partyId, layoutSetName, cancellationToken);
-            List<string> tasks = await _previewService.GetTasksForAllLayoutSets(org, app, developer, cancellationToken);
-            AppProcessState processState = new AppProcessState(mockInstance.Process)
-            {
-                ProcessTasks = tasks != null
-                    ? new List<AppProcessTaskTypeInfo>(tasks?.ConvertAll(task => new AppProcessTaskTypeInfo { ElementId = task, AltinnTaskType = "data" }))
-                    : null
-            };
-
-            return Ok(processState);
-        }
-
-        /// <summary>
-        /// Endpoint to get instance for next process step
-        /// </summary>
-        /// <returns>A mocked instance object</returns>
-        [HttpGet]
-        [Route("instances/{partyId}/{instanceGuId}")]
-        public async Task<ActionResult<Instance>> InstanceForNextTask(string org, string app, [FromRoute] int partyId, CancellationToken cancellationToken)
-        {
-            string developer = AuthenticationHelper.GetDeveloperUserName(_httpContextAccessor.HttpContext);
-            string refererHeader = Request.Headers["Referer"];
-            string layoutSetName = GetSelectedLayoutSetInEditorFromRefererHeader(refererHeader);
-            Instance mockInstance = await _previewService.GetMockInstance(org, app, developer, partyId, layoutSetName, cancellationToken);
-            return Ok(mockInstance);
-        }
-
-        /// <summary>
-        /// Endpoint to get active instances for apps with state/layout sets/multiple processes
-        /// </summary>
-        /// <returns>A list of a single mocked instance</returns>
-        [HttpGet]
-        [Route("instances/{partyId}/active")]
-        public ActionResult<List<Instance>> ActiveInstancesForAppsWithLayoutSets(string org, string app, [FromRoute] int partyId)
-        {
-            // Simulate never having any active instances
-            List<Instance> activeInstances = new();
-            return Ok(activeInstances);
-        }
-
-        /// <summary>
-        /// Endpoint to validate an instance
-        /// </summary>
-        /// <returns>Ok</returns>
-        [HttpGet]
-        [Route("instances/{partyId}/{instanceGuId}/validate")]
-        public ActionResult ValidateInstance()
-        {
-            return Ok();
-        }
-
-        /// <summary>
-        /// Endpoint to validate a data task for an instance
-        /// </summary>
-        /// <returns>Ok</returns>
-        [HttpGet]
-        [Route("instances/{partyId}/{instanceGuId}/data/" + PreviewService.MockDataTaskId + "/validate")]
-        public ActionResult ValidateInstanceForDataTask()
-        {
-            return Ok(new List<string>());
-        }
-
-        /// <summary>
-        /// Action for getting a mocked response for the next task connected to the instance
-        /// </summary>
-        /// <returns>The processState object on the global mockInstance object</returns>
-        [HttpGet]
-        [Route("instances/{partyId}/{instanceGuId}/process/next")]
-        public async Task<ActionResult> ProcessNext(string org, string app, [FromRoute] int partyId, CancellationToken cancellationToken)
-        {
-            string developer = AuthenticationHelper.GetDeveloperUserName(_httpContextAccessor.HttpContext);
-            string refererHeader = Request.Headers["Referer"];
-            string layoutSetName = GetSelectedLayoutSetInEditorFromRefererHeader(refererHeader);
-            Instance mockInstance = await _previewService.GetMockInstance(org, app, developer, partyId, layoutSetName, cancellationToken);
-            return Ok(mockInstance.Process);
-        }
-
-        /// <summary>
-        /// Action for mocking an end to the process in order to get receipt after "send inn" is pressed
-        /// </summary>
-        /// <returns>Process object where ended is set</returns>
-        [HttpPut]
-        [Route("instances/{partyId}/{instanceGuId}/process/next")]
-        public async Task<ActionResult> UpdateProcessNext(string org, string app, [FromRoute] int partyId, [FromQuery] string lang, CancellationToken cancellationToken)
-        {
-            string refererHeader = Request.Headers["Referer"];
-            string layoutSetName = GetSelectedLayoutSetInEditorFromRefererHeader(refererHeader);
-            if (string.IsNullOrEmpty(layoutSetName))
-            {
-                string endProcess = """{"ended": "ended"}""";
-                return Ok(endProcess);
-            }
-            string developer = AuthenticationHelper.GetDeveloperUserName(_httpContextAccessor.HttpContext);
-            Instance mockInstance = await _previewService.GetMockInstance(org, app, developer, partyId, layoutSetName, cancellationToken);
-            return Ok(mockInstance.Process);
         }
 
         /// <summary>
@@ -864,70 +628,6 @@ namespace Altinn.Studio.Designer.Controllers
                 // Return empty list since app-frontend don't handle a null result
                 return Ok(new List<string>());
             }
-        }
-
-        /// <summary>
-        /// Action for getting options list for a given options list id for a given instance
-        /// </summary>
-        /// <param name="org">Unique identifier of the organisation responsible for the app.</param>
-        /// <param name="app">Application identifier which is unique within an organisation.</param>
-        /// <param name="optionListId">The id of the options list</param>
-        /// <param name="language">The language for the options list</param>
-        /// <param name="source">The source of the options list</param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/> that observes if operation is cancelled.</param>
-        /// <returns>The options list if it exists, otherwise nothing</returns>
-        [HttpGet]
-        [Route("instances/{partyId}/{instanceGuid}/options/{optionListId}")]
-        public async Task<ActionResult<string>> GetOptionsForInstance(string org, string app, string optionListId, [FromQuery] string language, [FromQuery] string source, CancellationToken cancellationToken)
-        {
-            try
-            {
-                // TODO: Need code to get dynamic options list based on language and source?
-                string developer = AuthenticationHelper.GetDeveloperUserName(HttpContext);
-                AltinnAppGitRepository altinnAppGitRepository = _altinnGitRepositoryFactory.GetAltinnAppGitRepository(org, app, developer);
-                string options = await altinnAppGitRepository.GetOptionsList(optionListId, cancellationToken);
-                return Ok(options);
-            }
-            catch (NotFoundException)
-            {
-                // Return empty list since app-frontend don't handle a null result
-                return Ok(new List<string>());
-            }
-        }
-
-        /// <summary>
-        /// Action for getting data list for a given data list id for a given instance
-        /// </summary>
-        /// <param name="org">Unique identifier of the organisation responsible for the app.</param>
-        /// <param name="app">Application identifier which is unique within an organisation.</param>
-        /// <param name="dataListId">The id of the data list</param>
-        /// <param name="language">The language for the data list</param>
-        /// <param name="size">The number of items to return</param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/> that observes if operation is cancelled.</param>
-        /// <returns>The options list if it exists, otherwise nothing</returns>
-        [HttpGet]
-        [Route("instances/{partyId}/{instanceGuid}/datalists/{dataListId}")]
-        public ActionResult<List<string>> GetDataListsForInstance(string org, string app, string dataListId, [FromQuery] string language, [FromQuery] string size, CancellationToken cancellationToken)
-        {
-            // TODO: Should look into whether we can get some actual data here, or if we can make an "informed" mock based on the setup.
-            // For now, we just return an empty list.
-            return Ok(new List<string>());
-        }
-
-        /// <summary>
-        /// Action for updating data model with tag for attachment component // TODO: Figure out what actually happens here
-        /// </summary>
-        /// <param name="org">Unique identifier of the organisation responsible for the app.</param>
-        /// <param name="app">Application identifier which is unique within an organisation.</param>
-        /// <param name="currentPage">Current page in running app</param>
-        /// <param name="layoutSetId">Current layout set in running app</param>
-        /// <param name="dataTypeId">Connected datatype for that process task</param>
-        /// <returns>The options list if it exists, otherwise nothing</returns>
-        [HttpPost]
-        [Route("instances/{partyId}/{instanceGuid}/pages/order")]
-        public IActionResult UpdateAttachmentWithTag(string org, string app, [FromQuery] string currentPage, [FromQuery] string layoutSetId, [FromQuery] string dataTypeId)
-        {
-            return Ok();
         }
 
         /// <summary>
